requirements.txt
Flask==2.2.2
Flask_Migrate==4.0.4
Flask_Script==2.0.6
flask_sqlalchemy==3.0.3

Werkzeug~=2.2.2
python-dateutil~=2.8.2
alembic~=1.9.3
SQLAlchemy~=2.0.2